requirements.txt
aiohttp>=3.9.5
aioresponses == 0.7.6
mashumaro>=3.13
pytest-aio==1.9.0
pytest-cov==5.0.0
syrupy == 4.6.1